@@ -1,6 +1,6 @@
 {
     "name": "padhie/twitch-api-bundle",
-    "version": "2.0.3",
+    "version": "2.0.4",
     "type": "library",
     "keywords": [
         "twitch",
@@ -18,7 +18,7 @@
         }
     ],
     "require": {
-        "php": "^7.4 || ^8",
+        "php": "^8.2",
         "ext-curl": "*",
         "ext-json": "*",
         "guzzlehttp/guzzle": "^7.4",
@@ -27,10 +27,11 @@
         "guzzlehttp/promises": "^1.5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7",
+        "phpunit/phpunit": "^9.6",
         "phpstan/phpstan": "^1.5",
         "friendsofphp/php-cs-fixer": "^3.6",
-        "maglnet/composer-require-checker": "^3.8"
+        "maglnet/composer-require-checker": "^3.8",
+        "rector/rector": "^0.18.10"
     },
     "autoload": {
         "psr-4": {
