@@ -10,13 +10,10 @@ use Padhie\TwitchApiBundle\Response\Goals\GetCreatorGoalsResponse;
 /**
  * Scope: -
  */
-final class GetCreatorGoalsRequest implements RequestInterface
+final readonly class GetCreatorGoalsRequest implements RequestInterface
 {
-    private string $broadcasterId;
-
-    public function __construct(string $broadcasterId)
+    public function __construct(private string $broadcasterId)
     {
-        $this->broadcasterId = $broadcasterId;
     }
 
     public function getMethod(): string
